Made CSCDigiToRawModule a global module
By converting CSCDigiToRaw to be const thread safe, the module
CSDDigiToRawModule could be converted to a global.
